Revert "feat: byte-addressable memory model (uint8_t* mem)"
This reverts commit 6a4b2c99105bc09457d6b36fb1bd643dcd4d50c1.

+++ test_gauntlet_runner.c
@@ -132,12 +132,12 @@ int main(int argc, char **argv) {
             wubu_mir_prog_t prog;
             if (hd_build_mir(src, &prog) != 0) { printf("build failed\n"); return 1; }
             /* Allocate mem like hd_run_prog does (JIT needs it for LOAD/STORE) */
-            uint8_t *mem_ptr = prog.mem;
+            int64_t *mem_ptr = prog.mem;
             if (mem_ptr == NULL) {
                 int64_t mem_hi = prog.total_mem;
                 if ((int64_t)(prog.next_vr_hi) - 1 > mem_hi) mem_hi = (int64_t)(prog.next_vr_hi) - 1;
                 int64_t mem_size = (mem_hi < 1) ? 1 : (mem_hi + 1);
-                mem_ptr = (uint8_t *)calloc((size_t)mem_size, sizeof(int64_t));
+                mem_ptr = (int64_t *)calloc((size_t)mem_size, sizeof(int64_t));
             }
             wubu_mir_prog_t prog_copy = prog;
             prog_copy.mem = mem_ptr;
